feat: booking-service completed

+++ booking-service/src/main/java/com/saloon/booking_service/dto/SaloonReport.java
@@ -9,7 +9,7 @@ public class SaloonReport {
 
     private String saloonName;
 
-    private Double totalEarnings;
+    private int totalEarnings;
 
     private Integer totalBookings;
